autonomous_company_v10/pyproject.toml
[project]
name = "autonomous-company"
version = "10.0.0"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "structlog>=24.4",
    "opentelemetry-api>=1.25",
    "opentelemetry-sdk>=1.25",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "aiosqlite>=0.20",
    "click>=8.1",
    "rich>=13.0",
    "claude-agent-sdk>=0.1.64",
    "fastapi>=0.136",
    "uvicorn[standard]",
    "sse-starlette>=3.3",
    "simpleeval>=1.0",
    "python-dotenv>=1.2",
    "typer>=0.24",
    "anyio>=4.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=9",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5",
    "freezegun>=1.5",
    "hypothesis>=6.100",
]

[project.scripts]
autonomous-company = "autonomous_company.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/autonomous_company"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
]
